requirements.txt
flask>=3.0.0
waitress>=3.0.0
python-dotenv>=1.0.0
playwright>=1.48.0
playwright-stealth>=2.0.0
